requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Settings management
pydantic-settings==2.4.0

# Supabase
supabase==2.7.4

# AI / Agent
google-generativeai==0.8.3

# Scraping
scrapling==0.2.9

# Google Maps
googlemaps==4.10.0

# HTTP client (used internally)
httpx==0.27.2
